rename main_amazon.py to main.py
-recompiled binary (ready production testing)
-ignore spec on .gitignore
