+++ .gitignore
@@ -17,3 +17,4 @@ launch.json
 *.json
 *.csv
 *.xlsx
+main.spec
