+++ Helper Files/main.py
@@ -1,0 +1,110 @@
+from parser_constants import EXPECTED_SALES_CHANNELS, AMAZON_KEYS, ETSY_KEYS
+from parser_utils import clean_phone_number, get_country_code
+from file_utils import get_output_dir, is_windows_machine
+from database import SQLAlchemyOrdersDB
+from parse_orders import ParseOrders
+from datetime import datetime
+import logging
+import sys
+import csv
+import os
+
+
+# GLOBAL VARIABLES
+TESTING = False
+SALES_CHANNEL = 'Etsy'
+SKIP_ETONAS_FLAG = False
+EXPECTED_SYS_ARGS = 4
+VBA_ERROR_ALERT = 'ERROR_CALL_DADDY'
+VBA_KEYERROR_ALERT = 'ERROR_IN_SOURCE_HEADERS'
+VBA_OK = 'EXPORTED_SUCCESSFULLY'
+
+if is_windows_machine():
+    # 334 C:\Coding\Ebay\Working\Backups\Amazon exports\Collected exports\export 2021.04.29 EU.txt
+    # 1879 C:\Coding\Ebay\Working\Backups\Amazon exports\Collected exports\Amazon EU 2021-08-19.txt
+    # C:\Coding\Ebay\Working\Backups\Amazon exports\Collected exports\export COM 2021.06.28 - MXN new curency.txt
+    
+    ORDERS_SOURCE_FILE = r'C:\Coding\Ebay\Working\Backups\Etsy\EtsySoldOrders2021-8.csv'
+    # ORDERS_SOURCE_FILE = r'C:\Coding\Ebay\Working\Backups\Amazon exports\Collected exports\Amazon EU 2021-08-19.txt'
+else:
+    ORDERS_SOURCE_FILE = r'/home/devyo/Coding/Git/Amazon Orders Parser/Amazon exports/Collected exports/run4.txt'
+
+# Logging config:
+log_path = os.path.join(get_output_dir(client_file=False), 'loading_amazon_orders.log')
+logging.basicConfig(handlers=[logging.FileHandler(log_path, 'a', 'utf-8')], level=logging.DEBUG)
+
+
+def get_cleaned_orders(source_file:str, sales_channel:str, proxy_keys:dict) -> list:
+    '''returns cleaned orders (as cleaned in clean_orders func) from source_file arg path'''
+    delimiter = ',' if sales_channel == 'Etsy' else '\t'
+    raw_orders = get_raw_orders(source_file, delimiter)
+    cleaned_orders = clean_orders(raw_orders, sales_channel, proxy_keys)
+    return cleaned_orders
+
+def get_raw_orders(source_file:str, delimiter:str) -> list:
+    '''returns raw orders as list of dicts for each order in txt source_file'''
+    with open(source_file, 'r', encoding='utf-8') as f:
+        source_contents = csv.DictReader(f, delimiter=delimiter)
+        raw_orders = [{header : value for header, value in row.items()} for row in source_contents]
+    return raw_orders
+
+def clean_orders(orders:list, sales_channel:str, proxy_keys:dict) -> list:
+    '''performs universal data cleaning for amazon and etsy raw orders data'''
+    for order in orders:
+        try:
+            if sales_channel == 'Etsy':
+                # transform etsy country (Lithuania) to country code (LT)
+                country = order[proxy_keys['ship-country']]
+                order[proxy_keys['ship-country']] = get_country_code(country)
+            else:
+                # fix phone numbers in amazon from '+1 210-728-4548 ext. 01071' to a more friendly version
+                order['buyer-phone-number'] = clean_phone_number(order['buyer-phone-number'])
+        except KeyError as e:
+            logging.critical(f'Failed while cleaning loaded orders. Last order: {order} Err: {e}')
+            print(VBA_KEYERROR_ALERT)
+            sys.exit()
+    return orders
+
+
+def parse_args(testing=False):
+    '''returns arguments passed from VBA or hardcoded test environment'''
+    if testing:
+        print('--- RUNNING IN TESTING MODE. Using hardcoded args---')
+        return ORDERS_SOURCE_FILE, SALES_CHANNEL, SKIP_ETONAS_FLAG
+
+    try:
+        assert len(sys.argv) == EXPECTED_SYS_ARGS, 'Unexpected number of sys.args passed'
+        source_fpath = sys.argv[1]
+        sales_channel = sys.argv[2]
+        skip_etonas = True if sys.argv[3] == 'True' else False
+        logging.info(f'Accepted sys args on launch: source_fpath: {source_fpath}; sales_channel: {sales_channel}; skip_etonas: {skip_etonas}. Whole sys.argv: {list(sys.argv)}')
+        assert sales_channel in EXPECTED_SALES_CHANNELS, f'Unexpected sales_channel value passed from VBA side: {sales_channel}'
+        return source_fpath, sales_channel, skip_etonas
+    except Exception as e:
+        print(VBA_ERROR_ALERT)
+        logging.critical(f'Error parsing arguments on script initialization in cmd. Arguments provided: {list(sys.argv)} Number Expected: {EXPECTED_SYS_ARGS}.')
+        sys.exit()
+
+def main():
+    '''Main function executing parsing of provided txt/csv file and outputing csv, xlsx files'''
+    logging.info(f'\n\n NEW RUN STARTING: {datetime.today().strftime("%Y.%m.%d %H:%M")}')    
+    source_fpath, sales_channel, skip_etonas = parse_args(testing=TESTING)
+    
+    # Define order dict keys to use
+    proxy_keys = ETSY_KEYS if sales_channel == 'Etsy' else AMAZON_KEYS
+
+    # Get cleaned source orders
+    cleaned_source_orders = get_cleaned_orders(source_fpath, sales_channel, proxy_keys)
+    
+    db_client = SQLAlchemyOrdersDB(cleaned_source_orders, source_fpath, sales_channel, proxy_keys, testing=TESTING)
+    new_orders = db_client.get_new_orders_only()
+    logging.info(f'Loaded file contains: {len(cleaned_source_orders)}. Further processing: {len(new_orders)} orders')
+
+    # Parse orders, export target files
+    ParseOrders(new_orders, db_client, proxy_keys, sales_channel).export_orders(testing=TESTING, skip_etonas=skip_etonas)
+    print(VBA_OK)
+    logging.info(f'\nRUN ENDED: {datetime.today().strftime("%Y.%m.%d %H:%M")}\n')
+
+
+if __name__ == "__main__":
+    main()
